fix: moved routes out of server.js into UserRoutes.js
