+++ src/server.js
@@ -8,9 +8,6 @@
 // - connections to file storage 
 
 const express = require("express");
-const { User } = require("./users/UserModel");
-const { generateJWT } = require("./utils/middleware/jwtMiddleware");
-const { validateUserAuth } = require("./utils/middleware/authMiddleware");
 const cors = require ("cors");
 
 const app = express();
@@ -36,42 +33,6 @@ app.get("/", (request, response) => {
 });
 
 // // Server app configuration is finished by this point 
-
-// Auth Routes
-app.post("/signup", async (request, response) => {
-	// check that a username and password are provided in request.body
-	let username = request.body.username;
-	let password = request.body.password;
-
-	if (!username || !password) {
-		response.status(400).json({
-			message:"Incorrect or missing sign-up credentials provided."
-		})
-	}
-
-	// make a user in the DB using the username an password
-	let newUser = await User.create({username: username, password: password});
-
-	// make a JWT based on the username and userID
-	let newJwt = generateJWT(newUser.id, newUser.username);
-
-	// return the JWT
-	response.json({
-		jwt: newJwt,
-		user: {
-			id: newUser.id,
-			username: newUser.username
-		}
-	});
-});
-
-// // Export the app so that other files can control when the server
-// // starts and stops 
-app.get("/protectedRoute", validateUserAuth, (request, response) => {
-	response.json({
-		message:"You can see protected content because you're signed in!"
-	})
-})
 
 module.exports = {
 	app
